+++ source/scalers/scaler.h
@@ -28,6 +28,6 @@ void upscale_160x144_to_212x160(uint16_t* restrict src, uint16_t* restrict dst);
 void upscale_160x144_to_212x144(uint16_t* restrict src, uint16_t* restrict dst);
 void upscale_160x144_to_240x160(uint16_t* restrict src, uint16_t* restrict dst);
 //for SMS
-void downscale_240x192to240x160(uint16_t* restrict src, uint16_t* restrict dst);
+void downscale_240x192to240x160(uint32_t* restrict src, uint32_t* restrict dst);
 
 #endif
